Fix dynamic creation of S.{diff,changes} and S.Bool.{rise,fall}.
These combinators if they are created in a step add themselves as
dependencies only at the end of the step. However they did not check
if the rank of the mutable they were added to had changed its rank
during the step (their rank would not be affected by this change since
they are not added as a dependency yet). This could lead in certain
cases in the dependency having a lower rank than the mutable they
depend on and hence a violation of their semantics.

Fixes #8.